add unscheduled_visit_values setting, which defaults to null but takes precedence over unscheduled_visit_pattern
