+++ src/onPreprocess/setXdomain/removeUnscheduledVisits.js
@@ -1,0 +1,14 @@
+import { set } from 'd3';
+
+export default function removeUnscheduledVisits() {
+    if (!this.config.unscheduled_visits) {
+        if (this.config.unscheduled_visit_values)
+            this.config.x.domain = this.config.x.domain.filter(
+                visit => this.config.unscheduled_visit_values.indexOf(visit) < 0
+            );
+        else if (this.config.unscheduled_visit_pattern)
+            this.config.x.domain = this.config.x.domain.filter(
+                visit => !this.config.unscheduled_visit_pattern.test(visit)
+            );
+    }
+}
